+++ src/console/completion/components/CompletionItem.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styled from "../../colorscheme/styled";
+import styled from "../../styles/styled";
 
 const Container = styled.li<{
   shown: boolean;
