+++ src/console/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import ColorSchemeProvider from "./colorscheme/providers";
+import ColorSchemeProvider from "./styles/providers";
 import { AppProvider } from "./app/provider";
 import App from "./App";
